refactoring , added src folder
moved into src the folders __mocks_, __tests__, auth, config,middleware,user, app.js
