+++ src/__mocks__/fileMock.js
