+++ src/__mocks__/styleMock.js
